+++ frigate/app.py
@@ -647,6 +647,13 @@ class FrigateApp:
 
         self.init_logger()
         logger.info(f"Starting Frigate ({VERSION})")
+
+        if not os.environ.get("I_PROMISE_I_WONT_MAKE_AN_ISSUE_ON_GITHUB"):
+            print(
+                "Frigate 0.14 UNSTABLE - not for public use at this time. Please use Frigate stable"
+            )
+            sys.exit(1)
+
         try:
             self.ensure_dirs()
             try:
