+++ web/src/components/Sidebar.tsx
@@ -1,49 +1,15 @@
 import { IconType } from "react-icons";
-import {
-  LuConstruction,
-  LuFileUp,
-  LuFilm,
-  LuLayoutDashboard,
-  LuVideo,
-} from "react-icons/lu";
 import { NavLink } from "react-router-dom";
 import { Sheet, SheetContent } from "@/components/ui/sheet";
 import Logo from "./Logo";
 import { ENV } from "@/env";
-
-const navbarLinks = [
-  {
-    id: 1,
-    icon: LuLayoutDashboard,
-    title: "Dashboard",
-    url: "/",
-  },
-  {
-    id: 2,
-    icon: LuVideo,
-    title: "Live",
-    url: "/live",
-  },
-  {
-    id: 3,
-    icon: LuFilm,
-    title: "History",
-    url: "/history",
-  },
-  {
-    id: 4,
-    icon: LuFileUp,
-    title: "Export",
-    url: "/export",
-  },
-  {
-    id: 5,
-    icon: LuConstruction,
-    title: "UI Playground",
-    url: "/playground",
-    dev: true,
-  },
-];
+import { navbarLinks } from "@/pages/site-navigation";
+import SettingsNavItems from "./settings/SettingsNavItems";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipTrigger,
+} from "@/components/ui/tooltip";
 
 function Sidebar({
   sheetOpen,
@@ -53,35 +19,34 @@ function Sidebar({
   setSheetOpen: (open: boolean) => void;
 }) {
   const sidebar = (
-    <aside className="sticky top-0 overflow-y-auto scrollbar-hidden py-4 lg:pt-0 flex flex-col ml-1 lg:w-56 gap-0">
-      {navbarLinks.map((item) => (
-        <SidebarItem
-          key={item.id}
-          Icon={item.icon}
-          title={item.title}
-          url={item.url}
-          dev={item.dev}
-          onClick={() => setSheetOpen(false)}
-        />
-      ))}
+    <aside className="w-[52px] z-10 h-screen sticky top-0 overflow-y-auto scrollbar-hidden py-4 flex flex-col justify-between">
+      <div className="w-full flex flex-col gap-0 items-center">
+        <Logo className="w-8 h-8 mb-6" />
+        {navbarLinks.map((item) => (
+          <SidebarItem
+            key={item.id}
+            Icon={item.icon}
+            title={item.title}
+            url={item.url}
+            dev={item.dev}
+            onClick={() => setSheetOpen(false)}
+          />
+        ))}
+      </div>
+      <SettingsNavItems className="hidden md:flex flex-col items-center" />
     </aside>
   );
 
   return (
     <>
       <div className="hidden md:block">{sidebar}</div>
-
       <Sheet
         open={sheetOpen}
         modal={false}
         onOpenChange={() => setSheetOpen(false)}
       >
-        <SheetContent side="left" className="w-[120px]">
-          <div className="w-full flex flex-row justify-center">
-            <div className="w-10">
-              <Logo />
-            </div>
-          </div>
+        <SheetContent side="left" className="w-[90px]">
+          <div className="w-full flex flex-row justify-center"></div>
           {sidebar}
         </SheetContent>
       </Sheet>
@@ -102,18 +67,26 @@ function SidebarItem({ Icon, title, url, dev, onClick }: SidebarItemProps) {
 
   return (
     shouldRender && (
-      <NavLink
-        to={url}
-        onClick={onClick}
-        className={({ isActive }) =>
-          `py-4 px-2 flex flex-col lg:flex-row items-center rounded-lg gap-2 lg:w-full hover:bg-border ${
-            isActive ? "font-bold bg-popover text-popover-foreground" : ""
-          }`
-        }
-      >
-        <Icon className="w-6 h-6 mr-1" />
-        <div className="text-sm text-center">{title}</div>
-      </NavLink>
+      <Tooltip>
+        <NavLink
+          to={url}
+          onClick={onClick}
+          className={({ isActive }) =>
+            `mx-[10px] mb-6 flex flex-col justify-center items-center rounded-lg ${
+              isActive
+                ? "font-bold text-white bg-primary"
+                : "text-muted-foreground bg-secondary"
+            }`
+          }
+        >
+          <TooltipTrigger>
+            <Icon className="w-5 h-5 m-[6px]" />
+          </TooltipTrigger>
+        </NavLink>
+        <TooltipContent side="right">
+          <p>{title}</p>
+        </TooltipContent>
+      </Tooltip>
     )
   );
 }
