+++ web/src/context/providers.tsx
@@ -3,6 +3,7 @@ import { ThemeProvider } from "@/context/theme-provider";
 import { RecoilRoot } from "recoil";
 import { ApiProvider } from "@/api";
 import { IconContext } from "react-icons";
+import { TooltipProvider } from "@/components/ui/tooltip";
 
 type TProvidersProps = {
   children: ReactNode;
@@ -13,9 +14,11 @@ function providers({ children }: TProvidersProps) {
     <RecoilRoot>
       <ApiProvider>
         <ThemeProvider defaultTheme="light" storageKey="frigate-ui-theme">
-          <IconContext.Provider value={{ size: "20" }}>
-            {children}
-          </IconContext.Provider>
+          <TooltipProvider>
+            <IconContext.Provider value={{ size: "20" }}>
+              {children}
+            </IconContext.Provider>
+          </TooltipProvider>
         </ThemeProvider>
       </ApiProvider>
     </RecoilRoot>
