+++ web/src/index.css
@@ -22,3 +22,8 @@
     @apply bg-background text-foreground;
   }
 }
+
+@font-face {
+  font-family: "Inter";
+  src: url("../fonts/Inter-VariableFont_slnt,wght.ttf");
+}
