+++ web/src/types/live.ts
@@ -1,0 +1,1 @@
+export type LivePlayerMode = "webrtc" | "mse" | "jsmpeg" | "debug";
